Reduce amount of boilerplate systems (#3)
Use a single system on which all synced components and resources are registered. This minimizes the amount of boilerplate code needed in order to use the editor as no additional systems need to be added.

The flip side of only using 1 system is that serialization is no longer running in parallel. But since it no longer has to depend on other systems, I think it should be able to run in parallel with the other game systems.